Move config files. Switch from Jest to Vitest and update tests.
