--- jest.config.js
@@ -1,14 +1,0 @@
-const nextJest = require('next/jest');
-
-const createJestConfig = nextJest({
-  dir: './',
-});
-
-/** @type {import('jest').Config} */
-const customJestConfig = {
-  moduleDirectories: ['node_modules', '<rootDir>/'],
-  testEnvironment: 'jest-environment-jsdom',
-  setupFilesAfterEnv: ['<rootDir>/jest.setup.js'],
-};
-
-module.exports = createJestConfig(customJestConfig);
