--- jest.setup.js
@@ -1,2 +1,0 @@
-import '@testing-library/jest-dom';
-import '@testing-library/jest-dom/extend-expect';
