+++ package.json
@@ -20,8 +20,10 @@
     "lint": "next lint",
     "prepare": "husky install",
     "start": "next start",
-    "test": "jest",
-    "test:update": "jest -u",
+    "test": "vitest run",
+    "test:coverage": "vitest run --coverage",
+    "test:update": "vitest --update",
+    "test:watch": "vitest",
     "types": "tsc -p tsconfig.json --noEmit"
   },
   "dependencies": {
@@ -41,24 +43,25 @@
     "tailwind-merge": "1.12.0"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "5.16.5",
     "@testing-library/react": "14.0.0",
-    "@types/jest": "29.5.0",
     "@types/node": "18.15.11",
     "@types/react": "18.0.35",
+    "@vitejs/plugin-react": "3.1.0",
+    "@vitest/coverage-c8": "0.30.1",
     "autoprefixer": "10.4.14",
     "eslint": "8.38.0",
     "eslint-config-next": "13.3.0",
     "eslint-config-prettier": "8.8.0",
     "husky": "8.0.3",
-    "jest": "29.5.0",
-    "jest-environment-jsdom": "29.5.0",
+    "jsdom": "21.1.1",
     "lint-staged": "13.2.1",
     "postcss": "8.4.21",
     "prettier": "2.8.7",
     "prettier-plugin-tailwindcss": "0.2.7",
     "tailwindcss": "3.3.1",
     "typescript": "5.0.4",
+    "vite-tsconfig-paths": "4.2.0",
+    "vitest": "0.30.1",
     "webpack": "5.79.0",
     "windy-radix-palette": "0.6.0"
   }
