+++ src/components/Articles/AboutMe.tsx
@@ -1,5 +1,5 @@
+import { personal } from '@content';
 import { faUser } from '@fortawesome/free-solid-svg-icons';
-import { personal } from 'contentlayer/generated';
 import React from 'react';
 import Prose from '../Prose/Prose';
 import { SectionHeading } from '../SectionHeading/SectionHeading';
