+++ src/components/Header/Header.test.tsx
@@ -1,9 +1,8 @@
-import { render } from '../../test-utils';
+import { render } from '@test-utils';
+import { expect, test } from 'vitest';
 import { Header } from './Header';
 
-describe('<Header />', () => {
-  test('Snapshot', async () => {
-    const { container } = render(<Header />);
-    expect(container).toMatchSnapshot();
-  });
+test('Snapshot', async () => {
+  const { container } = render(<Header />);
+  expect(container).toMatchSnapshot();
 });
