+++ src/helpers/utils.ts
@@ -1,4 +1,4 @@
-import { personal } from 'contentlayer/generated';
+import { personal } from '@content';
 
 export const fullName = `${personal.givenName} ${personal.familyName}`;
